# API/DTOs/UserDto.cs
namespace API.DTOs
{
    public class UserDto
    {
        public string DisplayName { get; set; } = string.Empty;
        public string Token { get; set; } = string.Empty;
        public string Username { get; set; } = string.Empty;
    }
}