# Tests.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Activities API",
    "version": "v1"
  },
  "paths": {
    "/api/activities": {
      "get": {
        "tags": [ "Activities" ],
        "operationId": "GetActivities",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ActivityDto" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [ "Activities" ],
        "operationId": "CreateActivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActivityFormDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityDto" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        },
        "security": [ { "Bearer": [] } ]
      }
    },
    "/api/activities/{id}": {
      "get": {
        "tags": [ "Activities" ],
        "operationId": "GetActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityDto" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "tags": [ "Activities" ],
        "operationId": "EditActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActivityFormDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityDto" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" }
        },
        "security": [ { "Bearer": [] } ]
      },
      "delete": {
        "tags": [ "Activities" ],
        "operationId": "DeleteActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" }
        },
        "security": [ { "Bearer": [] } ]
      }
    },
    "/api/account/login": {
      "post": {
        "tags": [ "Account" ],
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/account/register": {
      "post": {
        "tags": [ "Account" ],
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/api/account": {
      "get": {
        "tags": [ "Account" ],
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [ { "Bearer": [] } ]
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityDto": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "title":       { "type": "string" },
          "date":        { "type": "string", "format": "date-time" },
          "description": { "type": "string" },
          "category":    { "type": "string" },
          "city":        { "type": "string" },
          "venue":       { "type": "string" }
        }
      },
      "ActivityFormDto": {
        "required": [ "title", "date", "description", "category", "city", "venue" ],
        "type": "object",
        "properties": {
          "title":       { "type": "string", "maxLength": 200 },
          "date":        { "type": "string", "format": "date-time" },
          "description": { "type": "string", "maxLength": 2000 },
          "category":    { "type": "string", "maxLength": 100 },
          "city":        { "type": "string", "maxLength": 100 },
          "venue":       { "type": "string", "maxLength": 200 }
        }
      },
      "LoginDto": {
        "required": [ "email", "password" ],
        "type": "object",
        "properties": {
          "email":    { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "RegisterDto": {
        "required": [ "displayName", "email", "password", "username" ],
        "type": "object",
        "properties": {
          "displayName": { "type": "string" },
          "email":       { "type": "string", "format": "email" },
          "password":    { "type": "string" },
          "username":    { "type": "string" }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "displayName": { "type": "string" },
          "token":       { "type": "string" },
          "username":    { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}